working but takes lot of time

+++ cpp/PixelWrapper.cpp
@@ -8,7 +8,7 @@
 
 #include "PixelWrapper.h"
 
-PixelWrapper::PixelWrapper(int x, int y, double value) {
+PixelWrapper::PixelWrapper(int x, int y, uchar value) {
     this->x = x;
     this->y = y;
     this->value = value;
